requirements.txt
python-dotenv>=0.14.0
RPi.bme280>=0.2.4
Flask>=2.2.3
matplotlib>=3.7.1
pandas>=1.4.0
waitress>=2.1.2
pysftp>=0.2.9
